alignment: filter observers based on the sun's altitude at the observer
Until now, potential observation points were filtered based on the
altitude of the sun at the target.  This means that the user had to guess
the altitude of the sun at the target corresponding to an altitude that
would put the sun behind the target from the observation point.  To
address this, the observation point is immediately computed instead of
being deferred until the filtering happened.  This new method also
takes refraction into account using Bennett/Sæmundsson formula.

As it is possible that Vincenty's formula cannot find an observation
point, the processing pipeline used to return the coordinates as an
option type.  They are now culled early in the pipeline, and all points
may be displayed.

Signed-off-by: Alban Gruin <alban@pa1ch.fr>