sundownjs: automatically compute the highest and lowest altitude bounds
This allows the web interface to compute automatically the highest and
lowest visual altitude bounds based on the distance between the observer
and the target, and the altitude of the observer and of the target.  The
bounds are automatically updated when one of those values are changed on
the interface.  They are found by computing the visual altitude of the
target, and by substrating (or adding, for the upper bound) half of the
sun's apparent diameter (fixed here to 0.52°).

The bounds are also no longer stored in presets.  Instead, the altitude
of the observer and of the target are saved.

The units are also now properly labelled on the interface.

Signed-off-by: Alban Gruin <alban@pa1ch.fr>