sundownjs: select the distance by simply clicking on the map
This allows to select the distance between the observer and the target
easily by simply clicking the map, after selecting the “distance” option
on the form.

This new field, as well as the previous “select target” field, are now
radio buttons.  As before, they are deselected when a click is
registered on the map, or when the form is submitted (i.e. alignments
are computed).

Signed-off-by: Alban Gruin <alban@pa1ch.fr>