alignment: more precise computation of the sunset time
This improves sunset computation by completing the implementation of
AA Chapter 14.

Additionally, this also opens the door to compute the sunrise or the
transit by parametrising the computation of `m' and its correction.
Ideally, it could be made fully generic by also parametrising the
computation of the body of interest and h0 (`stdh').

Signed-off-by: Alban Gruin <alban@pa1ch.fr>